updated to SDK 1.104
git-svn-id: svn+ssh://svn.code.sf.net/p/rts-2/code/fliusb@11038 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ libfli/unix/libfli-debug.c
@@ -114,7 +114,7 @@ void debug(int level, char *format, ...)
   return;
 }
 
-void setdebuglevel(char *host, int level)
+void setdebuglevel(char *host, long level)
 {
   _loghost = host;
   _loglevel = level;
